3.16.22
Everything is mostly working. there is still an issue with getting the different messages to display based on certain conditions so i am likely going to have to add in another div or header element maybe and make that the message target. but the score, timer, question generator, all are working at this point. Even though they are overly verbose.